BadgeView.java优化

+++ common/src/main/java/com/like/common/view/toolbar/ICustomToolbarMenu.kt
@@ -56,7 +56,7 @@ interface ICustomToolbarMenu {
     fun setMessageCount(
         messageCount: Int,
         @ColorInt textColor: Int? = null,
-        textSize: Int? = null,
+        textSize: Float? = null,
         @ColorInt backgroundColor: Int? = null
     )
